LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/fill-and-stroke-styles/2d.gradient.radial.cone.shape1-expected.txt

2d.gradient.radial.cone.shape1


FAIL OffscreenCanvas test: 2d.gradient.radial.cone.shape1 assert_equals: Green channel of the pixel at (98, 25) expected 255 but got 254
